+++ Rawdataproj2/WebService/Models/Search/BestmatchModel.cs
@@ -8,6 +8,6 @@ namespace WebService.Models.Search
     public class BestmatchModel
     {
         public int Id { get; set; }
-        public int Rank { get; set; }
+        public decimal Rank { get; set; }
     }
 }
